html_error_msg: Move this into its own file, to be easier to import
Specifically, so that it can be imported without django settings being configured.

This will be used by lib-cove-bods
