+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- Make `html_error_msg` importable from its own file, and without Django settings needing to be configured (e.g. in lib-cove-bods).
+
 # [0.18.4] - 2020-12-18
 
 ### Fixed
